+++ projects/ng2-panzoom/src/lib/panzoom-model.ts
@@ -1,4 +1,4 @@
-import { Point } from './panzoom-point';
+import { Point } from './types/point';
 
 export interface PanZoomModel {
   zoomLevel: number;