+++ projects/ng2-panzoom/src/lib/types/offset.ts
@@ -1,0 +1,4 @@
+export interface Offset {
+  top: number;
+  left: number;
+}